Pull chat composer out of chat app.
