+++ frontend/src/common/components/Chat.jsx
@@ -1,13 +1,9 @@
-import React, { createRef, createContext, useMemo } from 'react'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faArrowRight } from '@fortawesome/free-solid-svg-icons'
+import React, { createContext, useMemo } from 'react'
 
 import { useCurrentAuthUser } from '../utils/auth'
 import { useGetChatData, useGetMessages, sendMessage } from '../utils/chat'
 import { MESSAGE_TYPE, UNKNOWN_USER } from '../utils/constants'
 import { Error } from './Errors'
-
-export const ChatContext = createContext()
 
 function getAuthorClasses(participants, myUserId) {
     return Object.keys(participants)
@@ -42,29 +38,6 @@ export function Message({ data, myUserId, participants }) {
     )
 }
 
-function ChatComposer({ chatId, myUserId }) {
-    const textRef = createRef()
-
-    const doSend = () => {
-        const message = textRef.current.value
-        if (!myUserId || !message) return
-        sendMessage(chatId, {
-            message,
-            from: myUserId,
-        })
-        textRef.current.value = ''
-    }
-
-    return (
-        <div className="ChatComposer">
-            <textarea ref={textRef} className="Input" />
-            <button type="button" onClick={doSend} className="Button Send">
-                <FontAwesomeIcon icon={faArrowRight} className="IconBefore" />
-            </button>
-        </div>
-    )
-}
-
 function ChatLoading() {
     return <p className="Centered">Loading chat...</p>
 }
@@ -83,22 +56,33 @@ function getChatDisplayError(chatId, chat, myUserId) {
     return null
 }
 
-export function ChatApp({ chatId, header, conversation }) {
+export const ChatContext = createContext()
+
+export function ChatApp({ chatId, header, conversation, composer }) {
     const authUser = useCurrentAuthUser()
     const myUserId = authUser?.uid
     const chat = useGetChatData(`${chatId}~${myUserId}`)
     const messages = useGetMessages(chatId)
-    const chatMemo = useMemo(() => ({ chat, messages, myUserId }))
 
     const displayError = getChatDisplayError(chatId, chat, myUserId)
     const isReady = chat.isLoaded && !displayError
+
+    const sendChatMessage = (message) => {
+        if (!myUserId || !message) return
+        sendMessage(chatId, {
+            message,
+            from: myUserId,
+        })
+    }
+
+    const chatMemo = useMemo(() => ({ chat, messages, myUserId, sendChatMessage }))
 
     const chatAppEl = (
         <ChatContext.Provider value={chatMemo}>
             <div className="ChatAppInner HideScroll">
                 <div className="ChatHeader">{header}</div>
                 <div className="ChatMessages">{conversation}</div>
-                <ChatComposer chatId={chatId} myUserId={myUserId} />
+                <div className="ChatComposer">{composer}</div>
             </div>
         </ChatContext.Provider>
     )
